@@ -1,12 +1,9 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
-namespace BaseDataValidatorLibrary.Classes
+namespace BaseDataValidatorLibrary.Classes;
+
+public class Annotate
 {
-    public class Annotate
-    {
-        public static bool ValidEmail(string emailAddress)
-        {
-            return new EmailAddressAttribute().IsValid(emailAddress);
-        }
-    }
+    public static bool ValidEmail(string emailAddress) 
+        => new EmailAddressAttribute().IsValid(emailAddress);
 }
