@@ -1,41 +1,39 @@
-﻿using System;
-using System.ComponentModel.DataAnnotations;
+﻿using System.ComponentModel.DataAnnotations;
 using System.Text.RegularExpressions;
 
-namespace BaseDataValidatorLibrary.CommonRules
-{
-    [AttributeUsage(AttributeTargets.Field | AttributeTargets.Property)]
-    public class PasswordCheckAttribute : ValidationAttribute
-    {
-        /// <summary>
-        ///  Override of <see cref="ValidationAttribute.IsValid(object)" />
-        /// </summary>
-        public override bool IsValid(object value)
-        {
-            var validPassword = false;
-            var reason = string.Empty;
-            var password = (value == null) ? string.Empty : value.ToString();
+namespace BaseDataValidatorLibrary.CommonRules;
 
-            if (string.IsNullOrWhiteSpace(password) || password.Length < 6)
+[AttributeUsage(AttributeTargets.Field | AttributeTargets.Property)]
+public class PasswordCheckAttribute : ValidationAttribute
+{
+    /// <summary>
+    ///  Override of <see cref="ValidationAttribute.IsValid(object)" />
+    /// </summary>
+    public override bool IsValid(object value)
+    {
+        var validPassword = false;
+        var reason = string.Empty;
+        var password = (value == null) ? string.Empty : value.ToString();
+
+        if (string.IsNullOrWhiteSpace(password) || password.Length < 6)
+        {
+            reason = "new password must be at least 6 characters long. ";
+        }
+        else
+        {
+            var pattern = new Regex("((?=.*\\d)(?=.*[a-z])(?=.*[A-Z])(?=.*[@#$%]).{6,20})");
+            if (!pattern.IsMatch(password))
             {
-                reason = "new password must be at least 6 characters long. ";
+                reason += "Your new password must contain at least 1 symbol character and number.";
             }
             else
             {
-                var pattern = new Regex("((?=.*\\d)(?=.*[a-z])(?=.*[A-Z])(?=.*[@#$%]).{6,20})");
-                if (!pattern.IsMatch(password))
-                {
-                    reason += "Your new password must contain at least 1 symbol character and number.";
-                }
-                else
-                {
-                    validPassword = true;
-                }
+                validPassword = true;
             }
-
-            return validPassword;
-
         }
 
+        return validPassword;
+
     }
+
 }
