@@ -1,28 +1,26 @@
-﻿using System;
-using System.ComponentModel.DataAnnotations;
+﻿using System.ComponentModel.DataAnnotations;
 
-namespace BaseDataValidatorLibrary.CommonRules
+namespace BaseDataValidatorLibrary.CommonRules;
+
+/// <summary>
+/// Validates a property of type Enum is assigned.
+/// For this to work number the elements from 1 to x,
+/// do not assign 0 to an element.
+/// </summary>
+[AttributeUsage(AttributeTargets.Field | AttributeTargets.Property)]
+public class RequiredEnumAttribute : RequiredAttribute
 {
     /// <summary>
-    /// Validates a property of type Enum is assigned.
-    /// For this to work number the elements from 1 to x,
-    /// do not assign 0 to an element.
+    ///  Override of <see cref="ValidationAttribute.IsValid(object)" />
     /// </summary>
-    [AttributeUsage(AttributeTargets.Field | AttributeTargets.Property)]
-    public class RequiredEnumAttribute : RequiredAttribute
+    public override bool IsValid(object sender)
     {
-        /// <summary>
-        ///  Override of <see cref="ValidationAttribute.IsValid(object)" />
-        /// </summary>
-        public override bool IsValid(object sender)
+        if (sender == null)
         {
-            if (sender == null)
-            {
-                return false;
-            }
-
-            var type = sender.GetType();
-            return type.IsEnum && Enum.IsDefined(type, sender); ;
+            return false;
         }
+
+        var type = sender.GetType();
+        return type.IsEnum && Enum.IsDefined(type, sender); ;
     }
 }
