@@ -1,28 +1,26 @@
-﻿using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
+﻿using System.ComponentModel.DataAnnotations;
 
-namespace BaseDataValidatorLibrary.Helpers
+namespace BaseDataValidatorLibrary.Helpers;
+
+public class EntityValidationResult
 {
-    public class EntityValidationResult
+    /// <summary>
+    /// List of validation errors
+    /// </summary>
+    public IList<ValidationResult> Errors { get; set; }
+
+    /// <summary>
+    /// Model has errors
+    /// </summary>
+    public bool HasError => Errors.Count > 0;
+    /// <summary>
+    /// Model is valid
+    /// </summary>
+    public bool IsValid => Errors.Count == 0;
+    public bool IsNotValid => Errors.Count > 0;
+
+    public EntityValidationResult(IList<ValidationResult> errors = null)
     {
-        /// <summary>
-        /// List of validation errors
-        /// </summary>
-        public IList<ValidationResult> Errors { get; set; }
-
-        /// <summary>
-        /// Model has errors
-        /// </summary>
-        public bool HasError => Errors.Count > 0;
-        /// <summary>
-        /// Model is valid
-        /// </summary>
-        public bool IsValid => Errors.Count == 0;
-        public bool IsNotValid => Errors.Count > 0;
-
-        public EntityValidationResult(IList<ValidationResult> errors = null)
-        {
-            Errors = errors ?? new List<ValidationResult>();
-        }
+        Errors = errors ?? new List<ValidationResult>();
     }
 }
