@@ -1,40 +1,38 @@
-﻿using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
+﻿using System.ComponentModel.DataAnnotations;
 
-namespace BaseDataValidatorLibrary.Helpers
+namespace BaseDataValidatorLibrary.Helpers;
+
+/// <summary>
+/// Provides an alternate to code in ValidatorLibrary.ValidationHelper.IsValidEntity and Model.Validate
+/// </summary>
+public static class SimpleValidator
 {
     /// <summary>
-    /// Provides an alternate to code in ValidatorLibrary.ValidationHelper.IsValidEntity and Model.Validate
+    /// Validate the model and return a response, which includes any validation messages and an IsValid bit.
     /// </summary>
-    public static class SimpleValidator
+    public static ValidationResponse Validate(object model)
     {
-        /// <summary>
-        /// Validate the model and return a response, which includes any validation messages and an IsValid bit.
-        /// </summary>
-        public static ValidationResponse Validate(object model)
+        var results = new List<ValidationResult>();
+        var context = new ValidationContext(model);
+
+        var isValid = Validator.TryValidateObject(model, context, results, true);
+
+        return new ValidationResponse()
         {
-            var results = new List<ValidationResult>();
-            var context = new ValidationContext(model);
+            IsValid = isValid,
+            Results = results
+        };
+    }
 
-            var isValid = Validator.TryValidateObject(model, context, results, true);
+    /// <summary>
+    /// Validate model
+    /// </summary>
+    /// <param name="model">Class instance to validate</param>
+    /// <returns></returns>
+    public static bool IsModelValid(object model)
+    {
+        var response = Validate(model);
 
-            return new ValidationResponse()
-            {
-                IsValid = isValid,
-                Results = results
-            };
-        }
-
-        /// <summary>
-        /// Validate model
-        /// </summary>
-        /// <param name="model">Class instance to validate</param>
-        /// <returns></returns>
-        public static bool IsModelValid(object model)
-        {
-            var response = Validate(model);
-
-            return response.IsValid;
-        }
+        return response.IsValid;
     }
 }
