@@ -1,55 +1,51 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
+﻿using System.Text;
 using System.Text.RegularExpressions;
 using BaseDataValidatorLibrary.Helpers;
 
-namespace BaseDataValidatorLibrary.LanguageExtensions
+namespace BaseDataValidatorLibrary.LanguageExtensions;
+
+public static class ValidatorExtensions
 {
-    public static class ValidatorExtensions
+
+    /// <summary>
+    /// List errors if any
+    /// </summary>
+    /// <param name="sender"><see cref="EntityValidationResult"/></param>
+    /// <returns>list of 0 to many error messages</returns>
+    public static List<string> ErrorMessagesList(this EntityValidationResult sender)
+    {
+        string RemoveSpaces(string item) 
+            => Regex.Replace(item, @"\s+", " ");
+            
+        return sender.Errors.Select(validationResult 
+            => RemoveSpaces(validationResult.ErrorMessage.SplitCamelCase())).ToList();
+    }
+
+    /// <summary>
+    /// Displays any error messages using Console.WriteLine for unit test
+    /// </summary>
+    /// <param name="sender"></param>
+    public static void ShowErrorMessages(this EntityValidationResult sender)
+    {
+            
+        sender.ErrorMessagesList().ForEach(Console.WriteLine);
+            
+    }
+
+    public static string ErrorText(this EntityValidationResult sender)
     {
 
-        /// <summary>
-        /// List errors if any
-        /// </summary>
-        /// <param name="sender"><see cref="EntityValidationResult"/></param>
-        /// <returns>list of 0 to many error messages</returns>
-        public static List<string> ErrorMessagesList(this EntityValidationResult sender)
+        if (sender.ErrorMessagesList().Count == 1)
         {
-            string RemoveSpaces(string item) 
-                => Regex.Replace(item, @"\s+", " ");
-            
-            return sender.Errors.Select(validationResult 
-                => RemoveSpaces(validationResult.ErrorMessage.SplitCamelCase())).ToList();
+            return sender.ErrorMessagesList().FirstOrDefault();
         }
-
-        /// <summary>
-        /// Displays any error messages using Console.WriteLine for unit test
-        /// </summary>
-        /// <param name="sender"></param>
-        public static void ShowErrorMessages(this EntityValidationResult sender)
+        else
         {
-            
-            sender.ErrorMessagesList().ForEach(Console.WriteLine);
-            
-        }
-
-        public static string ErrorText(this EntityValidationResult sender)
-        {
-
-            if (sender.ErrorMessagesList().Count == 1)
-            {
-                return sender.ErrorMessagesList().FirstOrDefault();
-            }
-            else
-            {
-                StringBuilder builder = new();
-                sender.ErrorMessagesList().ForEach(x => builder.AppendLine(x));
-                return builder.ToString();
-            }
-
+            StringBuilder builder = new();
+            sender.ErrorMessagesList().ForEach(x => builder.AppendLine(x));
+            return builder.ToString();
         }
 
     }
+
 }
