@@ -1,51 +1,48 @@
-﻿using System;
-using System.ComponentModel;
+﻿using System.ComponentModel;
 using System.ComponentModel.DataAnnotations;
 using System.Runtime.CompilerServices;
 using BaseDataValidatorLibrary.CommonRules;
 
-namespace EntityFrameworkCoreLibrary.Models
+namespace EntityFrameworkCoreLibrary.Models;
+
+public class Customer : INotifyPropertyChanged
 {
-    public class Customer : INotifyPropertyChanged
+    private DateTime _birthDate;
+
+    [Key]
+    public int Id { get; set; }
+
+    [Required(ErrorMessage = "{0} is required"), DataType(DataType.Text)]
+    [MaxLength(12, ErrorMessage = "The {0} can not have more than {1} characters")]
+    [Display(Name = "First name")]
+    public string FirstName { get; set; }
+
+    [Required(ErrorMessage = "{0} is required"), DataType(DataType.Text)]
+    [MaxLength(12, ErrorMessage = "The {0} can not have more than {1} characters")]
+    [Display(Name = "Last name")]
+    public string LastName { get; set; }
+
+    [Required]
+    [YearRange(maximumYear: 2022, MinimumYear = 1932)]
+    [Display(Name = "Date of birth", Prompt = "Enter your birth date")]
+    public DateTime BirthDate
     {
-        private DateTime _birthDate;
-
-        [Key]
-        public int Id { get; set; }
-
-        [Required(ErrorMessage = "{0} is required"), DataType(DataType.Text)]
-        [MaxLength(12, ErrorMessage = "The {0} can not have more than {1} characters")]
-        [Display(Name = "First name")]
-        public string FirstName { get; set; }
-
-        [Required(ErrorMessage = "{0} is required"), DataType(DataType.Text)]
-        [MaxLength(12, ErrorMessage = "The {0} can not have more than {1} characters")]
-        [Display(Name = "Last name")]
-        public string LastName { get; set; }
-
-        [Required]
-        [YearRange(maximumYear: 2022, MinimumYear = 1932)]
-        [Display(Name = "Date of birth", Prompt = "Enter your birth date")]
-        public DateTime BirthDate
+        get => _birthDate;
+        set
         {
-            get => _birthDate;
-            set
-            {
-                _birthDate = value;
-                OnPropertyChanged();
-            }
+            _birthDate = value;
+            OnPropertyChanged();
         }
-
-
-
-        public override string ToString() => $"{FirstName} {LastName}";
-
-
-        public event PropertyChangedEventHandler PropertyChanged;
-        protected virtual void OnPropertyChanged([CallerMemberName] string propertyName = null)
-        {
-            PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(propertyName));
-        }
-
     }
+
+
+
+    public override string ToString() => $"{FirstName} {LastName}";
+    
+    public event PropertyChangedEventHandler PropertyChanged;
+    protected virtual void OnPropertyChanged([CallerMemberName] string propertyName = null)
+    {
+        PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(propertyName));
+    }
+
 }
