@@ -1,50 +1,48 @@
-﻿using System.Linq;
-using System.Text;
+﻿using System.Text;
 
-namespace ValidatorLibrary.Classes
+namespace ValidatorLibrary.Classes;
+
+/// <summary>
+/// Wrappers for model validation
+/// </summary>
+public class ValidationHelper
 {
     /// <summary>
-    /// Wrappers for model validation
+    /// Validate entity against validation rules
     /// </summary>
-    public class ValidationHelper
+    /// <typeparam name="T">model type</typeparam>
+    /// <param name="entity">instance of model</param>
+    /// <returns><see cref="EntityValidationResult"/> for <typeparam name="T">model type</typeparam></returns>
+    public static EntityValidationResult ValidateEntity<T>(T entity) where T : class 
+        => new EntityValidator<T>().Validate(entity);
+
+    /// <summary>
+    /// Validate entity
+    /// </summary>
+    /// <typeparam name="T">model</typeparam>
+    /// <param name="entity">Customer instance</param>
+    /// <returns>success and if not valid error messages</returns>
+    public static (bool success, string errorMessages) IsValidEntity<T>(T entity) where T : class
     {
-        /// <summary>
-        /// Validate entity against validation rules
-        /// </summary>
-        /// <typeparam name="T">model type</typeparam>
-        /// <param name="entity">instance of model</param>
-        /// <returns><see cref="EntityValidationResult"/> for <typeparam name="T">model type</typeparam></returns>
-        public static EntityValidationResult ValidateEntity<T>(T entity) where T : class 
-            => new EntityValidator<T>().Validate(entity);
+        var result = ValidateEntity(entity);
 
-        /// <summary>
-        /// Validate entity
-        /// </summary>
-        /// <typeparam name="T">model</typeparam>
-        /// <param name="entity">Customer instance</param>
-        /// <returns>success and if not valid error messages</returns>
-        public static (bool success, string errorMessages) IsValidEntity<T>(T entity) where T : class
+        if (result.IsNotValid)
         {
-            var result = ValidateEntity(entity);
-
-            if (result.IsNotValid)
+            StringBuilder builder = new();
+            if (result.Errors.Count == 1)
             {
-                StringBuilder builder = new();
-                if (result.Errors.Count == 1)
-                {
-                    return (false, result.Errors.FirstOrDefault().ErrorMessage);
-                }
-                else
-                {
-                    result.Errors.ToList().ForEach(x => builder.AppendLine(x.ErrorMessage));
-                    return (false, builder.ToString());
-                }
+                return (false, result.Errors.FirstOrDefault().ErrorMessage);
             }
             else
             {
-                return (true, null);
+                result.Errors.ToList().ForEach(x => builder.AppendLine(x.ErrorMessage));
+                return (false, builder.ToString());
             }
         }
-
+        else
+        {
+            return (true, null);
+        }
     }
+
 }
