@@ -1,41 +1,39 @@
 ﻿using System.ComponentModel;
-using System.Windows.Forms;
 
-namespace WindowsFormsLibrary.Controls
+namespace WindowsFormsLibrary.Controls;
+
+/// <summary>
+/// TextBox with no beep when ENTER is pressed as is is limited usage.
+/// </summary>
+public class NoBeepTextBox : TextBox
 {
+    public delegate void TriggerDelegate();
     /// <summary>
-    /// TextBox with no beep when ENTER is pressed as is is limited usage.
+    /// Subscribe to be notified when ENTER was pressed.
     /// </summary>
-    public class NoBeepTextBox : TextBox
+    public event TriggerDelegate TriggerEvent;
+    protected override void OnKeyDown(KeyEventArgs e)
     {
-        public delegate void TriggerDelegate();
-        /// <summary>
-        /// Subscribe to be notified when ENTER was pressed.
-        /// </summary>
-        public event TriggerDelegate TriggerEvent;
-        protected override void OnKeyDown(KeyEventArgs e)
+        if (e.KeyCode == (Keys.Return))
         {
-            if (e.KeyCode == (Keys.Return))
-            {
-                e.Handled = true;
-                e.SuppressKeyPress = true;
+            e.Handled = true;
+            e.SuppressKeyPress = true;
 
-                TriggerEvent?.Invoke();
+            TriggerEvent?.Invoke();
 
-                return;
-            }
-
-            base.OnKeyDown(e);
+            return;
         }
 
-        [Category("Behavior"), Description("Identifier")]
-        public int Id { get; set; }
-        [Browsable(false), EditorBrowsable(EditorBrowsableState.Never)]
-        public bool HasId => Id > -1;
-
-        [Category("Behavior"), Description("Stash")]
-        public string Stash { get; set; }
-
-
+        base.OnKeyDown(e);
     }
+
+    [Category("Behavior"), Description("Identifier")]
+    public int Id { get; set; }
+    [Browsable(false), EditorBrowsable(EditorBrowsableState.Never)]
+    public bool HasId => Id > -1;
+
+    [Category("Behavior"), Description("Stash")]
+    public string Stash { get; set; }
+
+
 }
