@@ -1,81 +1,78 @@
-﻿using System;
-using System.ComponentModel;
-using System.Windows.Forms;
+﻿using System.ComponentModel;
 
-namespace WindowsFormsLibrary.Controls
+namespace WindowsFormsLibrary.Controls;
+
+public class NullableDateTimePicker : DateTimePicker
 {
-    public class NullableDateTimePicker : DateTimePicker
+    private DateTimePickerFormat _originalFormat = DateTimePickerFormat.Short;
+    private string _originalCustomFormat;
+    private bool _isNull;
+
+    [Browsable(false), EditorBrowsable(EditorBrowsableState.Never)]
+    public new DateTime Value
     {
-        private DateTimePickerFormat originalFormat = DateTimePickerFormat.Short;
-        private string originalCustomFormat;
-        private bool isNull;
-
-        [Browsable(false), EditorBrowsable(EditorBrowsableState.Never)]
-        public new DateTime Value
+        get => _isNull ? DateTime.MinValue : base.Value;
+        set
         {
-            get => isNull ? DateTime.MinValue : base.Value;
-            set
+            // incoming value is set to min date
+            if (value == DateTime.MinValue)
             {
-                // incoming value is set to min date
-                if (value == DateTime.MinValue)
+                // if set to min and not previously null, preserve original formatting
+                if (!_isNull)
                 {
-                    // if set to min and not previously null, preserve original formatting
-                    if (!isNull)
-                    {
-                        originalFormat = Format;
-                        originalCustomFormat = CustomFormat;
-                        isNull = true;
-                    }
-
-                    Format = DateTimePickerFormat.Custom;
-                    CustomFormat = " ";
+                    _originalFormat = Format;
+                    _originalCustomFormat = CustomFormat;
+                    _isNull = true;
                 }
-                else // incoming value is real date
+
+                Format = DateTimePickerFormat.Custom;
+                CustomFormat = " ";
+            }
+            else // incoming value is real date
+            {
+                // if set to real date and previously null, restore original formatting
+                if (_isNull)
                 {
-                    // if set to real date and previously null, restore original formatting
-                    if (isNull)
-                    {
-                        Format = originalFormat;
-                        CustomFormat = originalCustomFormat;
-                        isNull = false;
-                    }
-
-                    base.Value = value;
+                    Format = _originalFormat;
+                    CustomFormat = _originalCustomFormat;
+                    _isNull = false;
                 }
+
+                base.Value = value;
             }
         }
-
-        protected override void OnCloseUp(EventArgs eventArgs)
-        {
-            // on keyboard close, restore format
-            if (MouseButtons == MouseButtons.None)
-            {
-                if (isNull)
-                {
-                    Format = originalFormat;
-                    CustomFormat = originalCustomFormat;
-                    isNull = false;
-                }
-            }
-
-            base.OnCloseUp(eventArgs);
-        }
-
-        protected override void OnKeyDown(KeyEventArgs e)
-        {
-            base.OnKeyDown(e);
-
-            // on delete key press, set to min value (null)
-            if (e.KeyCode == Keys.Delete)
-            {
-                Value = DateTime.MinValue;
-            }
-        }
-
-        /// <summary>
-        /// Indicates if Value has a date or not
-        /// </summary>
-        [Browsable(false), EditorBrowsable(EditorBrowsableState.Never)]
-        public bool IsNullValue => Value == DateTime.MinValue;
     }
+
+    protected override void OnCloseUp(EventArgs eventArgs)
+    {
+        // on keyboard close, restore format
+        if (MouseButtons == MouseButtons.None)
+        {
+            if (_isNull)
+            {
+                Format = _originalFormat;
+                CustomFormat = _originalCustomFormat;
+                _isNull = false;
+            }
+        }
+
+        base.OnCloseUp(eventArgs);
+    }
+
+    protected override void OnKeyDown(KeyEventArgs e)
+    {
+        base.OnKeyDown(e);
+
+        // on delete key press, set to min value (null)
+        if (e.KeyCode == Keys.Delete)
+        {
+            Value = DateTime.MinValue;
+        }
+    }
+
+    /// <summary>
+    /// Indicates if Value has a date or not
+    /// </summary>
+    [Browsable(false), EditorBrowsable(EditorBrowsableState.Never)]
+    public bool IsNullValue => Value == DateTime.MinValue;
 }
