@@ -1,9 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using EntityFrameworkCoreLibrary.Models;
+﻿using EntityFrameworkCoreLibrary.Models;
 using Microsoft.EntityFrameworkCore;
 
 namespace EntityFrameworkCoreLibrary.Data
@@ -11,11 +6,15 @@ namespace EntityFrameworkCoreLibrary.Data
     public class ExampleContext : DbContext
     {
         public DbSet<Customer> Customers { get; set; }
-        private static string _connectionString = "Server=(localdb)\\mssqllocaldb;Database=EF.DataAnnotations;Trusted_Connection=True";
+
+        /// <summary>
+        /// For production place in appsettings.json
+        /// </summary>
+        private static readonly string ConnectionString = "Server=(localdb)\\mssqllocaldb;Database=EF.DataAnnotations;Trusted_Connection=True";
 
         protected override void OnConfiguring(DbContextOptionsBuilder optionsBuilder)
         {
-            optionsBuilder.UseSqlServer(_connectionString);
+            optionsBuilder.UseSqlServer(ConnectionString);
         }
     }
 }
