@@ -1,8 +1,5 @@
-﻿using System;
-using System.Drawing;
-using System.Runtime.InteropServices;
+﻿using System.Runtime.InteropServices;
 using System.Text;
-using System.Windows.Forms;
 
 namespace WindowsFormsLibrary.Classes
 {
