@@ -1,19 +1,16 @@
-﻿using System.Windows.Forms;
+﻿namespace WindowsFormsLibrary.Controls;
 
-namespace WindowsFormsLibrary.Controls
+public class PasswordTextBox : TextBox
 {
-    public class PasswordTextBox : TextBox
+    public void TextMasked(bool show)
     {
-        public void TextMasked(bool show)
-        {
-            PasswordChar = show ?
-                '\0' :
-                '\u25CF';
-        }
+        PasswordChar = show ?
+            '\0' :
+            '\u25CF';
+    }
 
-        public PasswordTextBox()
-        {
-            TextMasked(false);
-        }
+    public PasswordTextBox()
+    {
+        TextMasked(false);
     }
 }
