@@ -1,31 +1,20 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Data;
-using System.Drawing;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows.Forms;
+﻿namespace WindowsFormsLibrary.Controls;
 
-namespace WindowsFormsLibrary.Controls
+public partial class PasswordTextBoxUserControl : UserControl
 {
-    public partial class PasswordTextBoxUserControl : UserControl
+    public PasswordTextBoxUserControl()
     {
-        public PasswordTextBoxUserControl()
-        {
-            InitializeComponent();
-        }
+        InitializeComponent();
+    }
 
-        public PasswordTextBox TextBox
-        {
-            get => PasswordTextBox;
-            set => PasswordTextBox = value;
-        }
+    public PasswordTextBox TextBox
+    {
+        get => PasswordTextBox;
+        set => PasswordTextBox = value;
+    }
 
-        private void ToggleCheckBox_CheckedChanged(object sender, EventArgs e)
-        {
-            PasswordTextBox.TextMasked(ToggleCheckBox.Checked);
-        }
+    private void ToggleCheckBox_CheckedChanged(object sender, EventArgs e)
+    {
+        PasswordTextBox.TextMasked(ToggleCheckBox.Checked);
     }
 }
